+++ packages/component/src/Styles/createStyleSet.ts
@@ -14,7 +14,6 @@ import createCarouselFilmStripAttachment from './StyleSet/CarouselFilmStripAttac
 import createCarouselFlipper from './StyleSet/CarouselFlipper';
 import createCitationModalDialogStyle from './StyleSet/CitationModalDialog';
 import createConnectivityNotification from './StyleSet/ConnectivityNotification';
-import createCSSCustomPropertiesStyle from './StyleSet/CSSCustomProperties';
 import createDictationInterimsStyle from './StyleSet/DictationInterims';
 import createErrorBoxStyle from './StyleSet/ErrorBox';
 import createErrorNotificationStyle from './StyleSet/ErrorNotification';
@@ -105,7 +104,6 @@ export default function createStyleSet(styleOptions: StyleOptions) {
     // Following styles follows new house rules:
     // - Use CSS var instead of strictStyleOptions
     citationModalDialog: createCitationModalDialogStyle(),
-    cssCustomProperties: createCSSCustomPropertiesStyle(strictStyleOptions),
     linkDefinitions: createLinkDefinitionsStyle(),
     modalDialog: createModalDialogStyle(),
     renderMarkdown: createRenderMarkdownStyle(),