+++ packages/component/src/Styles/createStyles.ts
@@ -1,0 +1,5 @@
+import { makeCreateStyles } from 'botframework-webchat-styles';
+
+export const componentStyleContent = '@--COMPONENT-STYLES-CONTENT--@';
+
+export default makeCreateStyles(componentStyleContent);