+++ packages/fluent-theme/src/styles/createStyles.ts
@@ -1,11 +1,5 @@
+import { makeCreateStyles } from 'botframework-webchat-styles';
+
 export const fluentStyleContent = '@--FLUENT-STYLES-CONTENT--@';
 
-export default function createStyles() {
-  if (!globalThis.document) {
-    return [];
-  }
-
-  const style = document.createElement('style');
-  style.append(document.createTextNode(fluentStyleContent));
-  return [style];
-}
+export default makeCreateStyles(fluentStyleContent);