+++ packages/fluent-theme/src/styles/useStyles.ts
@@ -1,19 +1,2 @@
-import { useMemo } from 'react';
-
-function useStyles<T extends CSSModuleClasses>(styles: T): T {
-  // @ts-expect-error: entries/fromEntries don't allow to specify keys type
-  return useMemo(
-    () =>
-      Object.freeze(
-        Object.fromEntries(
-          Object.entries(styles).map(([baseClassName, resultClassName]) => [
-            baseClassName,
-            `${baseClassName} ${resultClassName}`
-          ])
-        )
-      ),
-    [styles]
-  );
-}
-
-export default useStyles;
+// TODO: refactor after initial approval gained
+export { useStyles as default } from 'botframework-webchat-styles/react';