+++ packages/fluent-theme/tsup.config.ts
@@ -1,8 +1,9 @@
-import { join } from 'path';
+import { join } from 'node:path';
 import { defineConfig } from 'tsup';
-import { fileURLToPath } from 'url';
+import { fileURLToPath } from 'node:url';
 import baseConfig from '../../tsup.base.config';
 import { fluentStyleContent as fluentStyleContentPlaceholder } from './src/styles/createStyles';
+import { injectCSSPlugin } from 'botframework-webchat-styles/build';
 
 const umdResolvePlugin = {
   name: 'umd-resolve',
@@ -26,19 +27,10 @@ const umdResolvePlugin = {
     build.onResolve({ filter: /^botframework-webchat-component\/internal$/u }, () => ({
       path: join(fileURLToPath(import.meta.url), '../src/external.umd/botframework-webchat-component/internal.ts')
     }));
-  }
-};
 
-const injectCSSPlugin = {
-  name: 'inject-css-plugin',
-  setup(build) {
-    build.onEnd(result => {
-      const js = result.outputFiles.find(f => f.path.match(/(\.js|\.mjs)$/u));
-      const css = result.outputFiles.find(f => f.path.match(/(\.css)$/u));
-      if (css && js?.text.includes(fluentStyleContentPlaceholder)) {
-        js.contents = Buffer.from(js.text.replace(`"${fluentStyleContentPlaceholder}"`, JSON.stringify(css.text)));
-      }
-    });
+    build.onResolve({ filter: /^botframework-webchat-component\/decorator$/u }, () => ({
+      path: join(fileURLToPath(import.meta.url), '../src/external.umd/botframework-webchat-component/decorator.ts')
+    }));
   }
 };
 
@@ -51,7 +43,7 @@ export default defineConfig([
       ...baseConfig.loader,
       '.css': 'local-css'
     },
-    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin],
+    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin({ stylesPlaceholder: fluentStyleContentPlaceholder })],
     format: ['cjs']
   },
   {
@@ -61,7 +53,7 @@ export default defineConfig([
       ...baseConfig.loader,
       '.css': 'local-css'
     },
-    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin],
+    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin({ stylesPlaceholder: fluentStyleContentPlaceholder })],
     format: ['esm']
   },
   {
@@ -71,7 +63,7 @@ export default defineConfig([
       ...baseConfig.loader,
       '.css': 'local-css'
     },
-    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin, umdResolvePlugin],
+    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin({ stylesPlaceholder: fluentStyleContentPlaceholder }), umdResolvePlugin],
     format: 'iife',
     outExtension() {
       return { js: '.js' };
@@ -84,7 +76,7 @@ export default defineConfig([
       ...baseConfig.loader,
       '.css': 'local-css'
     },
-    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin, umdResolvePlugin],
+    esbuildPlugins: [...(baseConfig.esbuildPlugins || []), injectCSSPlugin({ stylesPlaceholder: fluentStyleContentPlaceholder }), umdResolvePlugin],
     format: 'iife',
     minify: true,
     outExtension() {